requirements.txt
chainlit
langgraph
langchain
langchain-community
langchain-google-genai
llama-index-core
llama-index-llms-openai
easyocr
pdf2image
pytesseract
sentence_transformers
python-dotenv
pymupdf
scikit-learn
pillow
torch
transformers